split component job features and steps
