+++ features/steps/jobs_steps.go
@@ -1,160 +1,17 @@
-// Package steps is used to define the steps that are used in the component test, which is written in godog (Go's version of cucumber).
 package steps
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"io"
-	"log"
-	"net/http"
-	"regexp"
 	"strconv"
-	"strings"
 	"time"
 
-	clientsidentity "github.com/ONSdigital/dp-api-clients-go/v2/identity"
-	clientssitesearch "github.com/ONSdigital/dp-api-clients-go/v2/site-search"
-	"github.com/ONSdigital/dp-authorisation/auth"
-	componentTest "github.com/ONSdigital/dp-component-test"
-	"github.com/ONSdigital/dp-component-test/utils"
-	"github.com/ONSdigital/dp-healthcheck/healthcheck"
-	kafka "github.com/ONSdigital/dp-kafka/v2"
-	"github.com/ONSdigital/dp-kafka/v2/kafkatest"
-	dpHTTP "github.com/ONSdigital/dp-net/v2/http"
-	"github.com/ONSdigital/dp-search-reindex-api/api"
-	"github.com/ONSdigital/dp-search-reindex-api/config"
-	"github.com/ONSdigital/dp-search-reindex-api/event"
 	"github.com/ONSdigital/dp-search-reindex-api/models"
-	"github.com/ONSdigital/dp-search-reindex-api/mongo"
-	"github.com/ONSdigital/dp-search-reindex-api/schema"
-	"github.com/ONSdigital/dp-search-reindex-api/service"
-	serviceMock "github.com/ONSdigital/dp-search-reindex-api/service/mock"
 	"github.com/cucumber/godog"
-	"github.com/pkg/errors"
 	"github.com/rdumont/assistdog"
-	uuid "github.com/satori/go.uuid"
 	"github.com/stretchr/testify/assert"
 )
-
-// collection names
-const jobsCol = "jobs"
-const tasksCol = "tasks"
-
-var (
-	// create map of valid task_name values for testing (in place of TaskNameValues in config)
-	taskName1, taskName2, taskName3, taskName4 = "dataset-api", "zebedee", "another-task-name3", "another-task-name4"
-
-	taskNames = map[string]bool{
-		taskName1: true,
-		taskName2: true,
-		taskName3: true,
-		taskName4: true,
-	}
-)
-
-// JobsFeature is a type that contains all the requirements for running a godog (cucumber) feature that tests the /jobs endpoint.
-type JobsFeature struct {
-	ErrorFeature            componentTest.ErrorFeature
-	svc                     *service.Service
-	errorChan               chan error
-	Config                  *config.Config
-	HTTPServer              *http.Server
-	ServiceRunning          bool
-	APIFeature              *componentTest.APIFeature
-	responseBody            []byte
-	MongoClient             *mongo.JobStore
-	MongoFeature            *componentTest.MongoFeature
-	AuthFeature             *componentTest.AuthorizationFeature
-	SearchFeature           *SearchFeature
-	KafkaProducer           service.KafkaProducer
-	MessageProducer         kafka.IProducer
-	quitReadingOutput       chan bool
-	createdJob              models.Job
-	kafkaProducerOutputData chan []byte
-}
-
-// NewJobsFeature returns a pointer to a new JobsFeature, which can then be used for testing the /jobs endpoint.
-func NewJobsFeature(mongoFeature *componentTest.MongoFeature,
-	authFeature *componentTest.AuthorizationFeature,
-	searchFeature *SearchFeature) (*JobsFeature, error) {
-	f := &JobsFeature{
-		HTTPServer:     &http.Server{},
-		errorChan:      make(chan error),
-		ServiceRunning: false,
-	}
-	svcErrors := make(chan error, 1)
-	cfg, err := config.Get()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get config: %w", err)
-	}
-	mongodb := &mongo.JobStore{
-		JobsCollection:  jobsCol,
-		TasksCollection: tasksCol,
-		Database:        utils.RandomDatabase(),
-		URI:             mongoFeature.Server.URI(),
-	}
-	ctx := context.Background()
-	if dbErr := mongodb.Init(ctx, cfg); dbErr != nil {
-		return nil, fmt.Errorf("failed to initialise mongo DB: %w", dbErr)
-	}
-
-	f.MongoClient = mongodb
-
-	f.AuthFeature = authFeature
-	cfg.ZebedeeURL = f.AuthFeature.FakeAuthService.ResolveURL("")
-
-	f.SearchFeature = searchFeature
-	cfg.SearchAPIURL = f.SearchFeature.FakeSearchAPI.ResolveURL("")
-
-	messageProducer := kafkatest.NewMessageProducer(true)
-	messageProducer.CheckerFunc = funcCheck
-	f.MessageProducer = messageProducer
-
-	kafkaProducer := &event.ReindexRequestedProducer{
-		Marshaller: schema.ReindexRequestedEvent,
-		Producer:   messageProducer,
-	}
-	f.KafkaProducer = kafkaProducer
-
-	f.kafkaProducerOutputData = make(chan []byte, 9999)
-	f.quitReadingOutput = make(chan bool)
-	f.readOutputMessages()
-
-	err = runJobsFeatureService(ctx, f, cfg, svcErrors)
-	if err != nil {
-		return nil, fmt.Errorf("failed to run JobsFeature service: %w", err)
-	}
-
-	return f, nil
-}
-
-// runJobsFeatureService uses the InitialiserMock to create the mock services that are required for testing the JobsFeature
-// It then uses these to create the external serviceList, which it passes into the service.Run function along with the fake identity and search clients.
-func runJobsFeatureService(ctx context.Context, f *JobsFeature, cfg *config.Config, svcErrors chan error) error {
-	var err error
-	initFunctions := &serviceMock.InitialiserMock{
-		DoGetHealthCheckFunc:           f.DoGetHealthcheckOk,
-		DoGetHTTPServerFunc:            f.DoGetHTTPServer,
-		DoGetMongoDBFunc:               f.DoGetMongoDB,
-		DoGetAuthorisationHandlersFunc: f.DoGetAuthorisationHandlers,
-		DoGetKafkaProducerFunc:         f.DoGetKafkaProducer,
-	}
-
-	serviceList := service.NewServiceList(initFunctions)
-	testIdentityClient := clientsidentity.New(cfg.ZebedeeURL)
-	testSearchClient := clientssitesearch.NewClient(cfg.SearchAPIURL)
-
-	f.svc, err = service.Run(ctx, cfg, serviceList, "1", "", "", svcErrors, testIdentityClient, taskNames, testSearchClient)
-	return err
-}
-
-// InitAPIFeature initialises the APIFeature that's contained within a specific JobsFeature.
-func (f *JobsFeature) InitAPIFeature() *componentTest.APIFeature {
-	f.APIFeature = componentTest.NewAPIFeature(f.InitialiseService)
-
-	return f.APIFeature
-}
 
 // RegisterSteps defines the steps within a specific JobsFeature cucumber test.
 func (f *JobsFeature) RegisterSteps(ctx *godog.ScenarioContext) {
@@ -209,83 +66,6 @@ func (f *JobsFeature) RegisterSteps(ctx *godog.ScenarioContext) {
 func (f *JobsFeature) iAmNotIdentifiedByZebedee() error {
 	f.AuthFeature.FakeAuthService.NewHandler().Get("/identity").Reply(401).BodyString(`{ "message": "user not authenticated"}`)
 	return nil
-}
-
-// Reset sets the resources within a specific JobsFeature back to their default values.
-func (f *JobsFeature) Reset(mongoFail bool) error {
-	if mongoFail {
-		f.MongoClient.Database = "lost database connection"
-	} else {
-		f.MongoClient.Database = utils.RandomDatabase()
-	}
-	if f.Config == nil {
-		cfg, err := config.Get()
-		if err != nil {
-			return fmt.Errorf("failed to get config: %w", err)
-		}
-		f.Config = cfg
-	}
-
-	return nil
-}
-
-// Close stops the *service.Service, which is pointed to from within the specific JobsFeature, from running.
-func (f *JobsFeature) Close() error {
-	close(f.quitReadingOutput)
-	if f.svc != nil && f.ServiceRunning {
-		err := f.svc.Close(context.Background())
-		if err != nil {
-			return fmt.Errorf("failed to close JobsFeature service: %w", err)
-		}
-		f.ServiceRunning = false
-	}
-	return nil
-}
-
-// InitialiseService returns the http.Handler that's contained within a specific JobsFeature.
-func (f *JobsFeature) InitialiseService() (http.Handler, error) {
-	return f.HTTPServer.Handler, nil
-}
-
-// DoGetHTTPServer takes a bind Address (string) and a router (http.Handler), which are used to set up an HTTPServer.
-// The HTTPServer is in a specific JobsFeature and is returned.
-func (f *JobsFeature) DoGetHTTPServer(bindAddr string, router http.Handler) service.HTTPServer {
-	f.HTTPServer.Addr = bindAddr
-	f.HTTPServer.Handler = router
-	return f.HTTPServer
-}
-
-// DoGetHealthcheckOk returns a mock HealthChecker service for a specific JobsFeature.
-func (f *JobsFeature) DoGetHealthcheckOk(cfg *config.Config, curTime, commit, version string) (service.HealthChecker, error) {
-	return &serviceMock.HealthCheckerMock{
-		AddCheckFunc: func(name string, checker healthcheck.Checker) error { return nil },
-		StartFunc:    func(ctx context.Context) {},
-		StopFunc:     func() {},
-	}, nil
-}
-
-// DoGetMongoDB returns a MongoDB, for the component test, which has a random database name and different URI to the one used by the API under test.
-func (f *JobsFeature) DoGetMongoDB(ctx context.Context, cfg *config.Config) (service.MongoDataStorer, error) {
-	return f.MongoClient, nil
-}
-
-// DoGetAuthorisationHandlers returns the mock AuthHandler that was created in the NewJobsFeature function.
-func (f *JobsFeature) DoGetAuthorisationHandlers(ctx context.Context, cfg *config.Config) api.AuthHandler {
-	authClient := auth.NewPermissionsClient(dpHTTP.NewClient())
-	authVerifier := auth.DefaultPermissionsVerifier()
-
-	// for checking caller permissions when we only have a user/service token
-	permissions := auth.NewHandler(
-		auth.NewPermissionsRequestBuilder(cfg.ZebedeeURL),
-		authClient,
-		authVerifier,
-	)
-	return permissions
-}
-
-// DoGetKafkaProducer returns a mock kafka producer.
-func (f *JobsFeature) DoGetKafkaProducer(ctx context.Context, cfg *config.Config) (service.KafkaProducer, error) {
-	return f.KafkaProducer, nil
 }
 
 // iWouldExpectJobIDLastupdatedAndLinksToHaveThisStructure is a feature step that can be defined for a specific JobsFeature.
@@ -436,18 +216,6 @@ func (f *JobsFeature) theResponseShouldContainAStateOf(expectedState string) err
 	assert.Equal(&f.ErrorFeature, expectedState, response.State)
 
 	return f.ErrorFeature.StepError()
-}
-
-// callPostJobs is a utility method that can be called by a feature step in order to call the POST /jobs endpoint
-// Calling that endpoint results in the creation of a job, in the Job Store, containing a unique id and default values.
-func (f *JobsFeature) callPostJobs() error {
-	var emptyBody = godog.DocString{}
-	err := f.APIFeature.IPostToWithBody("/jobs", &emptyBody)
-	if err != nil {
-		return fmt.Errorf("error occurred in IPostToWithBody: %w", err)
-	}
-
-	return nil
 }
 
 // iCallGETJobsidUsingTheGeneratedID is a feature step that can be defined for a specific JobsFeature.
@@ -988,186 +756,5 @@ func (f *JobsFeature) theReindexrequestedEventShouldContainTheExpectedJobIDAndSe
 	}
 	assert.Equal(&f.ErrorFeature, f.createdJob.ID, reindexRequestedData.JobID)
 	assert.Equal(&f.ErrorFeature, f.createdJob.SearchIndexName, reindexRequestedData.SearchIndex)
-	return nil
-}
-
-// GetJobByID is a utility function that is used for calling the GET /jobs/{id} endpoint.
-// It checks that the id string is a valid UUID before calling the endpoint.
-func (f *JobsFeature) GetJobByID(id string) error {
-	_, err := uuid.FromString(id)
-	if err != nil {
-		return fmt.Errorf("the id should be a uuid: %w", err)
-	}
-
-	// call GET /jobs/{id}
-	err = f.APIFeature.IGet("/jobs/" + id)
-	if err != nil {
-		return fmt.Errorf("error occurred in IGet: %w", err)
-	}
-	return nil
-}
-
-// PutNumberOfTasks is a utility function that is used for calling the PUT /jobs/{id}/number_of_tasks/{count}
-// It checks that the id string is a valid UUID before calling the endpoint.
-func (f *JobsFeature) PutNumberOfTasks(countStr string) error {
-	var emptyBody = godog.DocString{}
-	_, err := uuid.FromString(f.createdJob.ID)
-	if err != nil {
-		return fmt.Errorf("the id should be a uuid: %w", err)
-	}
-
-	// call PUT /jobs/{id}/number_of_tasks/{count}
-	err = f.APIFeature.IPut("/jobs/"+f.createdJob.ID+"/number_of_tasks/"+countStr, &emptyBody)
-	if err != nil {
-		return fmt.Errorf("error occurred in IPut: %w", err)
-	}
-	return nil
-}
-
-// PostTaskForJob is a utility function that is used for calling POST /jobs/{id}/tasks
-// The endpoint requires authorisation and a request body.
-func (f *JobsFeature) PostTaskForJob(jobID string, requestBody *godog.DocString) error {
-	_, err := uuid.FromString(jobID)
-	if err != nil {
-		return fmt.Errorf("the id should be a uuid: %w", err)
-	}
-
-	// call POST /jobs/{id}/tasks
-	err = f.APIFeature.IPostToWithBody("/jobs/"+jobID+"/tasks", requestBody)
-	if err != nil {
-		return fmt.Errorf("error occurred in IPostToWithBody: %w", err)
-	}
-	return nil
-}
-
-// GetTaskForJob is a utility function that is used for calling GET /jobs/{id}/tasks/{task name}
-func (f *JobsFeature) GetTaskForJob(jobID, taskName string) error {
-	_, err := uuid.FromString(jobID)
-	if err != nil {
-		return fmt.Errorf("the job id should be a uuid: %w", err)
-	}
-
-	// call GET /jobs/{jobID}/tasks/{taskName}
-	err = f.APIFeature.IGet("/jobs/" + jobID + "/tasks/" + taskName)
-	if err != nil {
-		return fmt.Errorf("error occurred in IPostToWithBody: %w", err)
-	}
-	return nil
-}
-
-// checkStructure is a utility function that can be called by a feature step to assert that a job contains the expected structure in its values of
-// id, last_updated, and links. It confirms that last_updated is a current or past time, and that the tasks and self links have the correct paths.
-func (f *JobsFeature) checkStructure(expectedResult map[string]string) error {
-	_, err := uuid.FromString(f.createdJob.ID)
-	if err != nil {
-		return fmt.Errorf("the id should be a uuid: %w", err)
-	}
-
-	if f.createdJob.LastUpdated.After(time.Now()) {
-		return errors.New("expected LastUpdated to be now or earlier but it was: " + f.createdJob.LastUpdated.String())
-	}
-
-	expectedLinksTasks := strings.Replace(expectedResult["links: tasks"], "{bind_address}", f.Config.BindAddr, 1)
-	expectedLinksTasks = strings.Replace(expectedLinksTasks, "{id}", f.createdJob.ID, 1)
-
-	assert.Equal(&f.ErrorFeature, expectedLinksTasks, f.createdJob.Links.Tasks)
-
-	expectedLinksSelf := strings.Replace(expectedResult["links: self"], "{bind_address}", f.Config.BindAddr, 1)
-	expectedLinksSelf = strings.Replace(expectedLinksSelf, "{id}", f.createdJob.ID, 1)
-
-	assert.Equal(&f.ErrorFeature, expectedLinksSelf, f.createdJob.Links.Self)
-
-	re := regexp.MustCompile(`(ons)(\d*)`)
-	wordWithExpectedPattern := re.FindString(f.createdJob.SearchIndexName)
-	assert.Equal(&f.ErrorFeature, wordWithExpectedPattern, f.createdJob.SearchIndexName)
-
-	return nil
-}
-
-// checkTaskStructure is a utility function that can be called by a feature step to assert that a job contains the expected structure in its values of
-// id, last_updated, and links. It confirms that last_updated is a current or past time, and that the tasks and self links have the correct paths.
-func (f *JobsFeature) checkTaskStructure(id string, lastUpdated time.Time, expectedResult map[string]string, links *models.TaskLinks, taskName string) error {
-	_, err := uuid.FromString(id)
-	if err != nil {
-		return fmt.Errorf("the jobID should be a uuid: %w", err)
-	}
-
-	if lastUpdated.After(time.Now()) {
-		return errors.New("expected LastUpdated to be now or earlier but it was: " + lastUpdated.String())
-	}
-
-	expectedLinksJob := strings.Replace(expectedResult["links: job"], "{bind_address}", f.Config.BindAddr, 1)
-	expectedLinksJob = strings.Replace(expectedLinksJob, "{id}", id, 1)
-
-	assert.Equal(&f.ErrorFeature, expectedLinksJob, links.Job)
-
-	expectedLinksSelf := strings.Replace(expectedResult["links: self"], "{bind_address}", f.Config.BindAddr, 1)
-	expectedLinksSelf = strings.Replace(expectedLinksSelf, "{id}", id, 1)
-	expectedLinksSelf = strings.Replace(expectedLinksSelf, "{task_name}", taskName, 1)
-
-	assert.Equal(&f.ErrorFeature, expectedLinksSelf, links.Self)
-	return nil
-}
-
-// checkValuesInJob is a utility function that can be called by a feature step in order to check that the values
-// of certain attributes, in a job, are all equal to the expected ones.
-func (f *JobsFeature) checkValuesInJob(expectedResult map[string]string, job models.Job) {
-	assert.Equal(&f.ErrorFeature, expectedResult["number_of_tasks"], strconv.Itoa(job.NumberOfTasks))
-	assert.Equal(&f.ErrorFeature, expectedResult["reindex_completed"], job.ReindexCompleted.Format(time.RFC3339))
-	assert.Equal(&f.ErrorFeature, expectedResult["reindex_failed"], job.ReindexFailed.Format(time.RFC3339))
-	assert.Equal(&f.ErrorFeature, expectedResult["reindex_started"], job.ReindexStarted.Format(time.RFC3339))
-	assert.Equal(&f.ErrorFeature, expectedResult["state"], job.State)
-	assert.Equal(&f.ErrorFeature, expectedResult["total_search_documents"], strconv.Itoa(job.TotalSearchDocuments))
-	assert.Equal(&f.ErrorFeature, expectedResult["total_inserted_search_documents"], strconv.Itoa(job.TotalInsertedSearchDocuments))
-}
-
-// checkValuesInTask is a utility function that can be called by a feature step in order to check that the values
-// of certain attributes, in a task, are all equal to the expected ones.
-func (f *JobsFeature) checkValuesInTask(expectedResult map[string]string, task models.Task) {
-	assert.Equal(&f.ErrorFeature, expectedResult["number_of_documents"], strconv.Itoa(task.NumberOfDocuments))
-	assert.Equal(&f.ErrorFeature, expectedResult["task_name"], task.TaskName)
-}
-
-// funcCheck is used to mock the kafka producer CheckerFunc
-func funcCheck(ctx context.Context, state *healthcheck.CheckState) error {
-	return nil
-}
-
-// readOutputMessages is a utility method to read the kafka messages that get sent to the producer's output channel
-func (f *JobsFeature) readOutputMessages() {
-	go func() {
-		for {
-			select {
-			case f.kafkaProducerOutputData <- <-f.MessageProducer.Channels().Output:
-				log.Println("read")
-			case <-f.quitReadingOutput:
-				return
-			}
-		}
-	}()
-}
-
-func readAndDeserializeKafkaProducerOutput(kafkaProducerOutputData <-chan []byte) (*models.ReindexRequested, error) {
-	for reindexRequestedDataBytes := range kafkaProducerOutputData {
-		reindexRequestedData := &models.ReindexRequested{}
-		err := schema.ReindexRequestedEvent.Unmarshal(reindexRequestedDataBytes, reindexRequestedData)
-		return reindexRequestedData, err
-	}
-	return nil, nil
-}
-
-// readResponse is a utility method to read the JSON response that gets returned by the POST /job endpoint
-func (f *JobsFeature) readResponse() error {
-	if (f.createdJob == models.Job{}) {
-		f.responseBody, _ = io.ReadAll(f.APIFeature.HttpResponse.Body)
-		response := models.Job{}
-		err := json.Unmarshal(f.responseBody, &response)
-		if err != nil {
-			return fmt.Errorf("failed to unmarshal json response: %w", err)
-		}
-
-		f.createdJob = response
-		return err
-	}
 	return nil
 }
